#13 cambio de variables de medicion de poblacion a uint32
Para asegurarse que no haya desbordamiento

+++ pkg/segurasenior/poblacion.go
@@ -4,16 +4,16 @@ import "errors"
 
 type Datos_poblacion struct {
 	Nombre_pueblo       string
-	Poblacion_total     uint16
-	Poblacion_joven     uint16
-	Poblacion_adulta    uint16
-	Poblacion_anciana   uint16
+	Poblacion_total     uint32
+	Poblacion_joven     uint32
+	Poblacion_adulta    uint32
+	Poblacion_anciana   uint32
 	Tasa_natalidad      float64
 	Tasa_envejecimiento float64
 	Tasa_mortalidad     float64
 }
 
-func NewPoblacion(nom string, poblacion uint16, poblacionjoven uint16, poblacionmedia uint16, poblacionmayor uint16, tasan float64, tasae float64, tasam float64) (*Datos_poblacion, error) {
+func NewPoblacion(nom string, poblacion uint32, poblacionjoven uint32, poblacionmedia uint32, poblacionmayor uint32, tasan float64, tasae float64, tasam float64) (*Datos_poblacion, error) {
 	if nom == "" {
 		return nil, errors.New("nombre no puede estar vacío")
 	}
